linkedin_agent/pyproject.toml
[project]
name = "linkedin-agent"
version = "0.1.0"
description = "LinkedIn content creation agent built with Google ADK"
requires-python = ">=3.9"
dependencies = [
    "google-adk",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "mypy",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true